Refactor: comm package unified, but there are errors. Pushing to test on a different platform since my workspace is messed up...
