+++ .vscode/dict.txt
@@ -7,6 +7,7 @@ BFT's
 Bitfield
 bitmask
 cerr
+cfgs
 chainedhotstuff
 checkf
 Chursin
