+++ protocol/consensus/voter_test.go
@@ -7,7 +7,7 @@ import (
 	"github.com/relab/hotstuff/internal/proto/clientpb"
 	"github.com/relab/hotstuff/internal/testutil"
 	"github.com/relab/hotstuff/protocol"
-	"github.com/relab/hotstuff/protocol/comm/clique"
+	"github.com/relab/hotstuff/protocol/comm"
 	"github.com/relab/hotstuff/protocol/consensus"
 	"github.com/relab/hotstuff/protocol/leaderrotation"
 	"github.com/relab/hotstuff/protocol/rules"
@@ -46,7 +46,7 @@ func wireUpVoter(
 		essentials.Authority(),
 		viewStates,
 	)
-	comm := clique.New(
+	comm := comm.NewClique(
 		essentials.RuntimeCfg(),
 		votingMachine,
 		leaderRotation,
